requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
sentence-transformers
transformers
torch
faiss-cpu
numpy
pypdf
PyPDF2
requests
huggingface_hub

pandas

